fix: bandwidth field type changed to 1nt64

+++ internal/models/manager.go
@@ -324,7 +324,7 @@ type QueueInfo struct {
 	Name            string
 	TargetDirectory string
 	MaxParallel     int
-	SpeedLimit      int
+	SpeedLimit      int64
 	NumRetries      int
 	StartTime       time.Time
 	EndTime         time.Time
